get rid of a compiler warning
[[Imported from SVN: r3889]]

+++ grid/test/checkindexset.cc
@@ -241,7 +241,7 @@ namespace Dune {
         std::cerr << "  " << *it << std::endl;
 
       std::cerr << std::endl << "but the method geomTypes() returned:" << std::endl;
-      for (int j=0; j<lset.geomTypes(codim).size(); j++)
+      for (size_t j=0; j<lset.geomTypes(codim).size(); j++)
         std::cerr << "  " << lset.geomTypes(codim)[j] << std::endl;
 
       DUNE_THROW(GridError, "!");
